+++ src/methods/payment/discount.rs
@@ -2,5 +2,5 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum DiscountValue {
-    Percentage(i128), Absolute(i128)
+    Percentage(i32), Absolute(i32)
 }
