+++ src/methods/product/product.rs
@@ -1,7 +1,8 @@
 use serde::{Serialize, Deserialize};
-use crate::methods::{Url, TagList, DiscountValue, TransactionType};
+use crate::{methods::{Url, TagList, DiscountValue}, entities::sea_orm_active_enums::TransactionType};
 use super::{VariantCategoryList, VariantIdTag};
 
+#[derive(Deserialize, Serialize)]
 pub struct Product {
     pub name: String,
     pub variants: VariantCategoryList,
@@ -22,8 +23,8 @@ pub struct ProductPurchase {
     pub variant: VariantIdTag,
     pub discount: DiscountValue,
 
-    pub product_cost: i128,
-    pub quantity: i128,
+    pub product_cost: i32,
+    pub quantity: i32,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -31,7 +32,7 @@ pub struct ProductExchange {
     pub method_type: TransactionType,
     pub product_code: ProductCode,
     pub variant: VariantIdTag,
-    pub quantity: i128,
+    pub quantity: i32,
 }
 
 pub type ProductCode = String;
