+++ src/methods/product/variant.rs
@@ -1,3 +1,5 @@
+use serde::{Deserialize, Serialize};
+
 use crate::methods::{StockList, HistoryList, Url};
 
 pub type VariantIdTag = Vec<VariantId>;
@@ -6,16 +8,18 @@ type VariantId = String;
 pub type VariantCategoryList = Vec<VariantCategory>;
 pub type VariantList = Vec<Variant>;
 
+#[derive(Deserialize, Serialize)]
 pub struct VariantCategory {
     pub category: String,
     pub variants: Variant
 }
 
+#[derive(Deserialize, Serialize)]
 pub struct Variant {
     pub name: String,
     pub stock: StockList,
     pub images: Vec<Url>,
-    pub marginal_price: i128,
+    pub marginal_price: i32,
     pub variant_code: String,
     pub order_history: HistoryList
 }
